chore(migrations): move 2.0.1 migrations into 2.0.0
2.0.1 is not a slotted release. 2.0.0 has not shipped yet, so the
in-progress 2.0.1 SQL and Java migrations are folded into 2.0.0 to
keep a single release boundary.

- SQL: append 2.0.1 schemaChanges.sql and postDataMigrationSQLScript.sql
  for mysql + postgres into the 2.0.0 equivalents; delete 2.0.1 dir.
- Java: rename v201 MigrationUtil to v200 TaskWorkflowMigrationUtil
  (avoids collision with existing v200 MigrationUtil) and wire its two
  public entrypoints (runTaskWorkflowCutoverMigration,
  addTaskResourceToMentionAlerts) into mysql/postgres v200 Migration
  classes after initializeWorkflowHandler().
- Tests: relocate v201 MigrationUtilTest to v200 TaskWorkflowMigrationUtilTest
  with the renamed class; 6 cases pass.
