+++ openmetadata-service/src/main/java/org/openmetadata/service/migration/mysql/v200/Migration.java
@@ -12,6 +12,7 @@ import static org.openmetadata.service.migration.utils.v200.MigrationUtil.migrat
 import lombok.SneakyThrows;
 import org.openmetadata.service.migration.api.MigrationProcessImpl;
 import org.openmetadata.service.migration.utils.MigrationFile;
+import org.openmetadata.service.migration.utils.v200.TaskWorkflowMigrationUtil;
 
 public class Migration extends MigrationProcessImpl {
 
@@ -35,5 +36,11 @@ public class Migration extends MigrationProcessImpl {
     backfillAnnouncementRelationships(handle);
     addTaskAuthorPolicyToDataConsumerRole(collectionDAO);
     addCreateTaskRuleToDataConsumerPolicy(collectionDAO);
+
+    initializeWorkflowHandler();
+    TaskWorkflowMigrationUtil taskWorkflowMigrationUtil = new TaskWorkflowMigrationUtil(handle);
+    taskWorkflowMigrationUtil.runTaskWorkflowCutoverMigration();
+    taskWorkflowMigrationUtil.addTaskResourceToMentionAlerts();
+    taskWorkflowMigrationUtil.runRecognizerFeedbackTaskTypeMigration();
   }
 }
