--- openmetadata-service/src/main/java/org/openmetadata/service/migration/postgres/v201/Migration.java
@@ -1,22 +1,0 @@
-package org.openmetadata.service.migration.postgres.v201;
-
-import lombok.SneakyThrows;
-import org.openmetadata.service.migration.api.MigrationProcessImpl;
-import org.openmetadata.service.migration.utils.MigrationFile;
-import org.openmetadata.service.migration.utils.v201.MigrationUtil;
-
-public class Migration extends MigrationProcessImpl {
-
-  public Migration(MigrationFile migrationFile) {
-    super(migrationFile);
-  }
-
-  @Override
-  @SneakyThrows
-  public void runDataMigration() {
-    initializeWorkflowHandler();
-    MigrationUtil migrationUtil = new MigrationUtil(handle);
-    migrationUtil.runTaskWorkflowCutoverMigration();
-    migrationUtil.addTaskResourceToMentionAlerts();
-  }
-}
